% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherInputFiles.R
\name{checkInputFiles}
\alias{checkInputFiles}
\title{Finding csv files with irregular format}
\usage{
checkInputFiles(inputPath, validationPath)
}
\arguments{
\item{inputPath}{Path of csv files obtained through Bluetooth Recoder}

\item{validationPath}{Path of csv files of validation files}
}
\description{
Finding csv files with irregular format
}
