% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherInputFiles.R
\name{gatherInputFiles}
\alias{gatherInputFiles}
\title{Reshaping csv files: put data of same validation form together}
\usage{
gatherInputFiles(inputPath, validationPath,
  outputFolderName = "step1_gather")
}
\arguments{
\item{inputPath}{Path of csv files obtained through Bluetooth Recoder}

\item{validationPath}{Path of csv files of validation files}

\item{outFolderName}{folder name for output data}
}
\description{
Reshaping csv files: put data of same validation form together
}
