% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReshapeData.R
\name{reshapeData}
\alias{reshapeData}
\title{Reshaping csv files in useful format}
\usage{
reshapeData(inputPath, outputFolderName = "step2_reshape")
}
\arguments{
\item{inputPath}{Path of csv files obtained through 'gatherInputFiles'}

\item{outFolderName}{Folder name for output data}
}
\description{
Reshaping csv files in useful format
}
