% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReshapeData.R
\name{reshapeDataEach}
\alias{reshapeDataEach}
\title{Reshaping csv files in useful format for each files}
\usage{
reshapeDataEach(inputPath, inputFileName, outputFolderName)
}
\arguments{
\item{inputPath}{Path of csv files obtained through 'gatherInputFiles'}

\item{inputFileName}{File Name of csv}
}
\description{
Reshaping csv files in useful format for each files
}
